// client/src/shared/api/generated/models/user-read.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenGraph API
 * AI/ML model and dataset Web3 blockchain infrastructure server
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * User read schema
 * @export
 * @interface UserRead
 */
export interface UserRead {
    /**
     * User email address
     * @type {string}
     * @memberof UserRead
     */
    'email': string;
    /**
     * 
     * @type {string}
     * @memberof UserRead
     */
    'google_id'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserRead
     */
    'display_name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserRead
     */
    'profile_image_url'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserRead
     */
    'sui_address'?: string | null;
    /**
     * User ID
     * @type {number}
     * @memberof UserRead
     */
    'id': number;
    /**
     * Creation timestamp
     * @type {string}
     * @memberof UserRead
     */
    'created_at': string;
}