// client/src/shared/api/generated/models/current-user-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenGraph API
 * AI/ML model and dataset Web3 blockchain infrastructure server
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 현재 사용자 응답
 * @export
 * @interface CurrentUserResponse
 */
export interface CurrentUserResponse {
    /**
     * 
     * @type {number}
     * @memberof CurrentUserResponse
     */
    'id': number;
    /**
     * 
     * @type {string}
     * @memberof CurrentUserResponse
     */
    'email': string;
    /**
     * 
     * @type {string}
     * @memberof CurrentUserResponse
     */
    'display_name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CurrentUserResponse
     */
    'profile_image_url'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CurrentUserResponse
     */
    'sui_address'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CurrentUserResponse
     */
    'google_id'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CurrentUserResponse
     */
    'zklogin_salt'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CurrentUserResponse
     */
    'created_at': string;
}